adecomposition of component to allow performance and dynamic loading
