+++ frontend/src/app/page/page.tsx
@@ -1,90 +1,43 @@
 "use client"
 
-import Image from 'next/image';
-import { useRef } from 'react';
 import FooterHome from '../components/footerHome';
 import HeaderHome from '../components/headerHomepage';
-import { Swiper, SwiperSlide } from 'swiper/react';
-import { useKeenSlider } from 'keen-slider/react'
-import 'swiper/css';
-import 'swiper/css/autoplay';
-import "keen-slider/keen-slider.min.css"
-import { Autoplay } from 'swiper/modules';
+import dynamic from 'next/dynamic';
+
+const Gallery = dynamic(() => import('../components/FarmerGallery'),{
+    ssr: false,
+    loading: () => (
+        <div className="flex items-center justify-center py-10">
+          <div className="w-12 h-12 border-4 border-green-500 border-t-transparent rounded-full animate-spin"></div>
+        </div>
+    ),
+})
+
+const Service = dynamic(() => import('../components/service'),{
+    ssr: false,
+    loading: () => (
+        <div className="flex items-center justify-center py-10">
+          <div className="w-12 h-12 border-4 border-green-500 border-t-transparent rounded-full animate-spin"></div>
+        </div>
+    ),
+})
+const PoweredTrusted = dynamic(() => import("../components/homepagepowered"), {
+    ssr: false,
+    loading: () => (
+        <div className="flex items-center justify-center py-10">
+          <div className="w-12 h-12 border-4 border-green-500 border-t-transparent rounded-full animate-spin"></div>
+        </div>
+    ),
+})
 
 export  default function Homepage() {
 
     const currentYear = new Date().getFullYear();
 
-    const serviceDiv ="bg-white rounded-xl shadow-md p-6 text-center hover:shadow-lg transition-shadow duration-300";
-    const Hdiv = "text-xl font-semibold text-green-800 mb-2";
-    const pservice = "text-gray-700 text-sm";
-    const ImageDiv = "w-80 h-50 relative mb-4 overflow-hidden rounded-md";
     const pFarmersSay = "text-gray-800 italic mb-4";
     const pFarmerName = "text-green-800 font-semibold text-sm";
 
     const Description ="This is a platform designed to help farmers track and improve productiivity in there farms by simple clicks where you enjoy seamless advantages of AI, ML prediction and timely production"
-
-    const logos =[
-        {src: '/logo.png',  alt: 'logo' },
-        {src: '/log1.jpeg',  alt: 'logo' },
-        {src: '/cow3.jpg',  alt: 'logo' },
-        {src: '/logo.png',  alt: 'logo' },
-        {src: '/logo.png',  alt: 'logo' },
-       
-    ]
-    const farmersImages = [
-        { id: 1, name: "crop framing", image: '/farmer6.jpg' },
-        { id: 2, name: "soil", image: '/farmer2.jpg' },
-        { id: 3, name: "crop farming", image: '/farmer3.jpg' },
-        { id: 4, name: "spinach farming", image: '/farmer4.jpg' },
-        { id: 6, name: "Irrigation", image: '/farmer5.jpg' },
-        { id: 5, name: "Watering", image: '/farmer1.jpg'}
-    ]
-    function Autoplay(slider:any){
-        let timeout: NodeJS.Timeout
-        let mouseOver = false;
-
-
-        function clearNextTimeout() {
-            clearTimeout(timeout)
-        }
-
-        function nextTimeout() {
-            clearTimeout(timeout)
-            if (mouseOver) return
-            timeout = setTimeout(() => {
-                slider.next()
-            }, 3000) 
-        }
-
-        slider.on("created", () => {
-            slider.container.addEventListener("mouseover", () => {
-                mouseOver = true
-                clearNextTimeout()
-            })
-        slider.container.addEventListener("mouseout", () => {
-            mouseOver = false
-            nextTimeout()
-        })
-        nextTimeout()
-        })
-        slider.on("dragStarted", clearNextTimeout)
-        slider.on("animationEnded", nextTimeout)
-        slider.on("updated", nextTimeout)
-    }
-
-    const [slider] = useKeenSlider<HTMLDivElement>({
-        loop: true,
-        mode: "free-snap",
-        slides:{
-            perView: 1.2,
-            spacing: 15,
-        },
-        breakpoints: {
-            "(min-width: 768px)": { slides: { perView: 2.5, spacing: 20 },},
-            "(min-width: 1024px)": { slides: {perView: 3.5, spacing: 25}, },
-        }
-    }, [Autoplay])
 
     return(
         <div className="min-h-screen flex flex-col">
@@ -98,6 +51,9 @@ export  default function Homepage() {
                         autoPlay 
                         muted 
                         loop
+                        preload="none"
+                        playsInline
+                        poster="/farmer5.jpg"
                         className="absolute top-0 left-0 w-full h-full object-cover z-0"
                     >
                         <source src="/background.mp4" type="video/mp4" />
@@ -114,89 +70,9 @@ export  default function Homepage() {
                         </p>
                     </div>
                 </section>
-                <section className="bg-green-50 py-12 px-4">
-                    <h3 className="text-center text-2xl text-green-700 font-bold mb-6">Farmers gallery</h3>
-                    <div ref={slider} className="keen-slider">
-                        {farmersImages.map((images) => (
-                            <div
-                                key={images.id}
-                                className="keen-slider__slide bg-white rounded-xl shadow-md overflow-hidden"
-                            >
-                                <Image 
-                                  src= {images.image} 
-                                  alt= {images.name}
-                                  width ={300}
-                                  height ={300}
-                                  className="w-full h-48 object-cover"
-                                />
-                                
-                            </div>
-                        ))}
-                    </div>
-                </section>
-
-                <section className="py-12 px-4 bg-green-100 ">
-                    <h2 className="text-3xl md:text-4xl font-bold text-center text-green-700 mb-12">services</h2>
-                    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-10 max-w-6xl mx-auto">
-                        <div className={serviceDiv}>
-                            <div className={ImageDiv}>
-                                <Image  src="/agri_cow.jpg" alt="Inventory" fill className="cover"/>
-                            </div>
-                            <h3 className={Hdiv}>Farm Inventory</h3>
-                            <p className={pservice}>You get your record well architectured and simplified for you to understand. The inventory helps you to plan for the farm management </p>
-                        </div>
-
-                        <div className={serviceDiv}>
-                            <div className={ImageDiv}>
-                                <Image  src="/Agri4.jpg" alt="AI image" fill className="cover"/>
-                            </div>
-                            <h3 className={Hdiv}>AI-powered Insights on your farm</h3>
-                            <p className={pservice}>Help analyze upto 97% correct prediction of the outcome of your produce and recommendation on how on steps to take to have a higher yield outcome</p>
-                            
-                        </div>
-
-                        <div className={serviceDiv}>
-                            <div className={ImageDiv}>
-                                <Image  src="/Agri5.png" alt="Real-analytics" fill className="cover"/>
-                            </div>
-                            <h3 className={Hdiv}>Real-time Farm Analytics</h3>
-                            <p className={pservice}>You get real-time updates on  your farm management and where to put more effort for higher productivity</p>
-                        </div>
-
-                        <div className={serviceDiv}>
-                            <div className={ImageDiv}>
-                                <Image  src="/Agri3.jpg" alt="Irrigation" fill className="cover"/>
-                            </div>
-                            <h3 className={Hdiv}>Smart Irrigation</h3>
-                            <p className={pservice}>The Platform is designed to help you as a farmer migrate  and get started with modern and emerging  Agricultural technology form irrifation to dairy farming and many more</p>
-                        </div>
-
-                        <div className={serviceDiv}>
-                            <div className={ImageDiv}>
-                                <Image  src="/cow3.jpg" alt="breeding" fill className="cover"/>
-                            </div>
-                            <h3 className={Hdiv}>Breeding Tracking</h3>
-                            <p className={pservice}>Track breeding cycles with up to 99% accuracy, ensuring optimal timing and better breed management.</p>
-                        </div>
-                        <div className={serviceDiv}>
-                            <div className={ImageDiv}>
-                                <Image src="/AgriDoc.png" alt="Animal health" fill className="cover"/>
-                            </div>
-                            <h3 className={Hdiv}>Farm health Insights</h3>
-                            <p className={pservice}> We offer recommendation for farms in controlling pesticide and how to fight animal diseases</p>
-
-                        </div>
-                        
-                    </div>
-                </section>
-
-                {/* <section>
-                    <h2>Farm Management Resources</h2>
-                    <div>
-
-                    </div>
-
-                </section> */}
+                <Gallery />
+                
+                <Service />
 
                 <section className="py-12 px-4 bg-white">
                     <h3 className="text-3xl md:text-4xl font-bold text-center text-green-700 mb-10">What Farmers are saying</h3>
@@ -215,44 +91,8 @@ export  default function Homepage() {
                         </div>
                     </div>
                 </section>
+                <PoweredTrusted/>
 
-                <section className="bg-green-100 py-12 px-4">
-                    <div className="relative z-10 text-center mb-12">
-                        <h2 className="text-3xl md:text-4xl font-bold text-green-800 mb-4">
-                           🤝 Trusted By & Powered By
-                        </h2>
-                        <p className="text-gray-600 max-w-xl mx-auto">
-                            Our technology is backed by industry leaders and trusted by farmers worldwide.
-                        </p>
-                    </div>
-                    <div className="max-w-6xl mx-auto">
-                        <Swiper 
-                          modules ={[Autoplay]}
-                          spaceBetween={30}
-                          slidesPerView={2}
-                          breakpoints ={{
-                            640: { slidesPerView: 3},
-                            768: { slidesPerView: 4},
-                            1024: { slidesPerView: 5},
-                          }}
-                          loop ={true}
-                          autoplay={{ delay: 2000, disableOnInteraction: false}}
-                        >
-                            {logos.map((logo, index)=> (
-                               <SwiperSlide key ={index}  className="flex justify-center items-center">
-                                  <Image 
-                                      src={logo.src}
-                                      alt={logo.alt}
-                                      width={90}
-                                      height={80}
-                                      className="object-contain grayscale hover:grayscale-0 transition duration-300 rounded-4xl"
-                                    />
-                               </SwiperSlide>
-                            ))}
-                        </Swiper>
-
-                    </div>
-                </section>
             </main>
             <footer>
                 <FooterHome />
